Refactored the ScheduleEntryDialog, removed the trigger from the dialog since the parent needs to be re-rendered when monthly schedule is fetched hence making the dialog to automatically close since the trigger is also being re-rendered.
