+++ src/components/features/scheduler/Scheduler.tsx
@@ -20,6 +20,7 @@ import mergeScheduleIntoCalendar from "@mr/lib/functions/merge-schedule-into-cal
 import { motion } from "framer-motion";
 import { Skeleton } from "@mr/components/ui/Skeleton";
 import { MonthYearPicker } from "../calendar/MonthYearPicker";
+import { ScheduleEntryDialog } from "./ScheduleEntryDialog";
 
 export const Scheduler: FunctionComponent = () => {
   const currentSchedule = useSchedulesStore((state) => state.currentSchedule);
@@ -339,6 +340,7 @@ export const Scheduler: FunctionComponent = () => {
           )}
         </main>
       </div>
+      <ScheduleEntryDialog />
     </>
   );
 };
